requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
requests==2.25.1
requests-futures==1.0.0
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.6
grequests~=0.6.0
PyMySQL~=1.0.2
tabulate~=0.8.9
python-dotenv~=0.19.0